requirements.txt
Flask_SocketIO==3.0.0
Flask==1.0.2
Pyserial==3.4
